feat(admin): implement admin dashboard

See merge request enigma-camp/oflline-class-lovelace/lovelace-17/001-tunduh/final-project/backend!9

+++ src/main/java/com/tunduh/timemanagement/service/impl/TaskServiceImpl.java
@@ -8,37 +8,47 @@ import com.tunduh.timemanagement.exception.ResourceNotFoundException;
 import com.tunduh.timemanagement.repository.TaskRepository;
 import com.tunduh.timemanagement.repository.UserRepository;
 import com.tunduh.timemanagement.service.TaskService;
+import com.tunduh.timemanagement.utils.pagination.CustomPagination;
+import com.tunduh.timemanagement.utils.specification.TaskSpecification;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class TaskServiceImpl implements TaskService {
-
     private final TaskRepository taskRepository;
     private final UserRepository userRepository;
 
     @Override
+    @Transactional
     public TaskResponse createTask(TaskRequest taskRequest, String userId) {
         UserEntity user = userRepository.findById(userId)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found"));
 
         TaskEntity task = TaskEntity.builder()
-                .id(String.valueOf(UUID.randomUUID()))
+                .id(UUID.randomUUID().toString())
                 .title(taskRequest.getTitle())
                 .energy(taskRequest.getEnergy())
-                .repetitionConfig(taskRequest.getRepetitionConfig())
                 .notes(taskRequest.getNotes())
                 .status(taskRequest.getStatus())
                 .duration(taskRequest.getDuration())
                 .priority(taskRequest.getPriority())
+                .repetitionType(TaskEntity.RepetitionType.valueOf(taskRequest.getRepetitionType()))
+                .repetitionDays(taskRequest.getRepetitionDays())
+                .repetitionEndDate(taskRequest.getRepetitionEndDate())
                 .user(user)
                 .createdAt(LocalDateTime.now())
                 .updatedAt(LocalDateTime.now())
@@ -49,38 +59,44 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    public List<TaskResponse> getAllTasks(String userId) {
-        List<TaskEntity> tasks = taskRepository.findByUserId(userId);
-        return tasks.stream()
+    public CustomPagination<TaskResponse> getAllTasks(String userId, int page, int size, String sort, String title, String status) {
+        Pageable pageable = createPageable(page, size, sort);
+        Specification<TaskEntity> spec = TaskSpecification.getSpecification(title, status, userId);
+
+        Page<TaskEntity> taskPage = taskRepository.findAll(spec, pageable);
+        List<TaskResponse> taskResponses = taskPage.getContent().stream()
                 .map(this::mapToTaskResponse)
-                .collect(Collectors.toList());
+                .toList();
+
+        return new CustomPagination<>(taskPage.map(this::mapToTaskResponse));
     }
 
     @Override
-    public TaskResponse getTaskById(String id) {
-        TaskEntity task = taskRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Task not found"));
+    public TaskResponse getTaskById(String id, String userId) {
+        TaskEntity task = taskRepository.findByIdAndUserId(id, userId);
+        if (task == null) {
+            throw new ResourceNotFoundException("Task not found for this user");
+        }
         return mapToTaskResponse(task);
     }
 
     @Override
     @Transactional
     public TaskResponse updateTask(String id, TaskRequest taskRequest, String userId) {
-        Optional<TaskEntity> optionalTask = Optional.ofNullable(taskRepository.findByIdAndUserId(id, userId));
-
-        if (optionalTask.isEmpty()) {
+        TaskEntity task = taskRepository.findByIdAndUserId(id, userId);
+        if (task == null) {
             throw new ResourceNotFoundException("Task not found for this user");
         }
 
-        TaskEntity task = optionalTask.get();
-
         task.setTitle(taskRequest.getTitle());
         task.setEnergy(taskRequest.getEnergy());
-        task.setRepetitionConfig(taskRequest.getRepetitionConfig());
         task.setNotes(taskRequest.getNotes());
         task.setStatus(taskRequest.getStatus());
         task.setDuration(taskRequest.getDuration());
         task.setPriority(taskRequest.getPriority());
+        task.setRepetitionType(TaskEntity.RepetitionType.valueOf(taskRequest.getRepetitionType()));
+        task.setRepetitionDays(taskRequest.getRepetitionDays());
+        task.setRepetitionEndDate(taskRequest.getRepetitionEndDate());
         task.setUpdatedAt(LocalDateTime.now());
 
         TaskEntity updatedTask = taskRepository.save(task);
@@ -91,6 +107,9 @@ public class TaskServiceImpl implements TaskService {
     @Transactional
     public void deleteTask(String id, String userId) {
         TaskEntity task = taskRepository.findByIdAndUserId(id, userId);
+        if (task == null) {
+            throw new ResourceNotFoundException("Task not found for this user");
+        }
         taskRepository.delete(task);
     }
 
@@ -99,13 +118,30 @@ public class TaskServiceImpl implements TaskService {
                 .id(UUID.fromString(task.getId()))
                 .title(task.getTitle())
                 .energy(task.getEnergy())
-                .repetitionConfig(task.getRepetitionConfig())
                 .notes(task.getNotes())
                 .status(task.getStatus())
                 .duration(task.getDuration())
                 .priority(task.getPriority())
+                .repetitionType(task.getRepetitionType())
+                .repetitionDays(task.getRepetitionDays())
+                .repetitionEndDate(task.getRepetitionEndDate())
                 .createdAt(task.getCreatedAt())
                 .updatedAt(task.getUpdatedAt())
                 .build();
+    }
+
+    private Pageable createPageable(int page, int size, String sort) {
+        List<Sort.Order> orders = new ArrayList<>();
+        if (sort != null) {
+            String[] sortParams = sort.split(",");
+            for (String param : sortParams) {
+                String[] keyDirection = param.split(":");
+                String key = keyDirection[0];
+                Sort.Direction direction = keyDirection.length > 1 && keyDirection[1].equalsIgnoreCase("desc") ?
+                        Sort.Direction.DESC : Sort.Direction.ASC;
+                orders.add(new Sort.Order(direction, key));
+            }
+        }
+        return PageRequest.of(page, size, Sort.by(orders));
     }
 }
